[Release 2.10.33-ex120] 现代化代码格式

+++ src/main/java/ic2/api/entity/block/ExplosiveEntity.java
@@ -167,7 +167,7 @@ public abstract class ExplosiveEntity extends Entity
 
 	public int getFuse()
 	{
-		return (Integer) this.entityData.get(FUSE);
+		return this.entityData.get(FUSE);
 	}
 
 	public void setFuse(int fuse)
