+++ src/main/java/ic2/core/block/generator/tileentity/TileEntityRTGenerator.java
@@ -17,7 +17,7 @@ public class TileEntityRTGenerator extends TileEntityBaseGenerator
 
 	public TileEntityRTGenerator(BlockPos pos, BlockState state)
 	{
-		super(Ic2BlockEntities.RT_GENERATOR, pos, state, Math.round(16.0F * (float) IC2Config.balance.energy.generator.radioisotope.get().floatValue()), 1, 20000);
+		super(Ic2BlockEntities.RT_GENERATOR, pos, state, Math.round(16.0F * IC2Config.balance.energy.generator.radioisotope.get().floatValue()), 1, 20000);
 		this.fuelSlot.setStackSizeLimit(1);
 	}
 
@@ -39,7 +39,7 @@ public class TileEntityRTGenerator extends TileEntityBaseGenerator
 			return false;
 		}
 
-		this.energy.addEnergy(Math.pow(2.0, counter - 1) * (float) IC2Config.balance.energy.generator.radioisotope.get().floatValue());
+		this.energy.addEnergy(Math.pow(2.0, counter - 1) * IC2Config.balance.energy.generator.radioisotope.get().floatValue());
 		return true;
 	}
